requirements.txt
annotated-types==0.6.0
anyio==4.3.0
autopep8==2.0.4
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
pycodestyle==2.11.1
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
tomli==2.0.1
typing_extensions==4.10.0
uvicorn==0.27.1
